Added workload generator and scripts to automatically bring up volumes, policies, and run workloads (on a single SH for now).
Workload generator is in tools/pharos. It has to be build separately
for now, since we are still investigating whether we will be using
it or not. It's currently does async IO or single threaded sync IO
accessing one of /dev/fbdN devices (device is passed as a command line option).

See tools/README for info and example how to use scripts to run workloads.