Firebreak event support
Fixes problem where firePostPersist() was calling prePersist, so the firebreak
event detection wasn't actually occurring.

Adds a feature toggle for the firebreak event handling as this is really an
in-progress feature that we might want to be able disable.

Refactored the HttpAuthenticator handling of the case where authentication is
disabled so that if the request passes authentication information it will use
it. If authentication is turned off, the request will execute regardless but
will ensure there is an AuthenticationToken (possibly ANONYMOUS) in the request.