Capture firebreak events
Refactor FirebreakHelper to expose helpers for the EventManager persist
listener;
Move the EventManager to om/events package since it references OM repository.
Add FirebreakEvent type

This also includes attempt to filter cases where there is already an active
firebreak instead of generating a new event.  This is not currently working.