Firebreak event support - continued.
Refactoring of the FirebreakHelper to allow the metric postPersist
handler access to firebreak information.  In testing I found a couple
of instances where the VolumeDatapointPair (previously just "Pair") had
incorrect data in it (wrong metrics), so I added validation when
creating the pair.  Similarly, I found that the ConsecutiveStream
is not guaranteed to give us pairs containing STC/LTC and STP/LTP values
so refactored that logic as well.