Updates to FirebreakEvent support - fix query issues.
Note that this includes changes to address Firebreak Event queries that
were returning the same results as previous queries for different
volumes.  There are changes included here in JDORepository handing of
the EntityManager that were initial and unsuccessful attempts at fixing that
There appears to be a bug in the ObjectDB EntityManager/CriteriaQuery
implementation. I think those changes with the handling of EntityManager
are the appropriate way to interact with an EntityManager in general.
The workaround for the issue with the query cache is to create a new
EventRepository object for each FirebreakEvent query.

Note that the same workaround may need to be applied to other queries
but that is not done in this commit.