AM Nbd: remove a bunch of ugly cast code for UIO.
Unfortunately the UIO is using a buffer for the kernel that's used for
reads and writes so the internal pointers can not be const. This caused
a bunch of ugly cast code which I've removed here generically. I could
have called this function "deconst" because it will work on anything
that's a pointer, but it be really bad to misuse.