Refactor XdiClientFactory to ThriftClientFactory
This extracts the base Thrift client functionality out of the XDI and
into a util package where it can be accessed by the OM and other
components as needed without introducing an explicit dependency on the
Xdi.  The XdiClientFactory class itself still lives in the Xdi, but now
uses the base ThriftClientFactory.

This work is in preparation for removing the dependency between the OM
and the Xdi to address issues with ConfigDB access and Events, in
addition to facilitating the multi-OM cluster communication.  Further
changes will be required of the XdiClientFactory, particularly having to
do with the use of the ConfigurationService thrift interface.