Refactor ConfigurationApi out of xdi
This is the first step towards separating the OM-specific configuration
API and the XDI configuration access.  First extracted the
com.formationds.util.thrift.ConfigurationApi interface (NOTE: not yet
convinced this is the appropriate place for this interface to live),
adding several read operations that are commonly used, usually through
the cache.  Then added an OmConfigurationApi implementation and renamed
the Xdi implementation to XdiConfigurationApi.  Modified all of the
references to use the interface.

This refactoring sets up the next stage which will involve modifying the
OM implementation to interact with the OM RAFT cluster; and the XDI
implementation to communicate with the OM REST Api for
configuration-related operations.