XDI to OM Rest Client for create/delete volume
Implemented a OM ConfigurationService Proxy that forwards create/delete
volume operations through the OM REST client.  All other
ConfigurationApi operations continue to go through the
XdiConfigurationApi to the OM ThriftServer running in the OM native
code.

In order to handle AuthenticationToken requirements when calling the
REST Api, I modified the S3Endpoint and SwiftAuthenticator to store the
token in the AuthenticatedRequestContext for the duration of the request
handling.  This is similar to how it is handled in the OM web request
handling.

Added debug output to the SmokeTest of the admin and user buckets to
allow us to track errors to the system log files.