queue stat stream registrations
Add StatStreamRegistrationHandler timed task to process queued stat
stream registration requests from create/delete volume operations.

This offloads the processing from the Create/Delete volume and queues
up the operations so they can be batched.  This especially helps in
situations where there are a large number of create/delete requests
coming in within a short period of time.  This is really only expected
in test scenarios, but we were encountering crashes that appeared to
be related to the frequent stream deregister/register that were occurring.

Note that this is partly necessary since the native OM and other components
don't currently support a mechanism to register/deregister of a single volume
at a time.  Once the native OM supports this, we can revisit this handler.