SM superblock knows about all SM tokens even if SM does not own them.
So that SM can handle 16-node cluster. The behavior is correct but not
very efficient. E.g., we will open levelDBs for SM tokens SM does not own,
or GC may try to run on those SM tokens (but will correctly find out that
there is nothing to GC). The more efficient implementation would be to
update SM superblock and SM state when SM gains or loses ownership of SM tokens.