fs-1113 - in-progress work NOT TESTED
Implemented first cut at the InfluxMetricRepository (event not yet there).

Continuing refactor metric/event repository interfaces to support a abstract factory de-coupled from the implementation of the underlying data storage mechanism.  Interfaces were necessary to implement the dynamic proxy approach to switching ObjectDB/InfluxDB based on the feature toggles we defined earlier.

It is possible we could have made this dynamic proxy approach work based on just the CRUDRepository interface, but it seemed like that would miss the metric or event specific api that is necessary.