fs-1418 OM hangs if influx is not running
This makes the InfluxDB connection execute in a retry loop and modifies the initial createDatabase to run asynchronously after the connection is available.  Write operations are also asynchronous once the connection is available, so the initial "Create Admin user" system event is handled without blocking.

This also fixes a query issue - the Where clause was not being added when querying just a time range.