fs-1114 Influx query support and related fixes
This completes the InfluxDB query support that is the minimum required to enable the QUERY_INFLUXDB feature toggle.

Remove generated id from Event.  It was not used and InfluxDB does not support generated id.
Change FirebreakEvent volume id to a long, which is consistent with Thrift definition of VolumeDescriptor as the event is isolated.
Added TODO comment to Volume model object that we need to make it consistent with the thrift VolumeDescriptor volId.
Add convenience constructors to DateRange, Volume, and QueryCriteria classes.
Update VolumeDatapointEntityPersistListener to use the event repository returned from the repository manager.

Complete implementation of influxdb queries for events.
Complete unimplemented methods for influxdb metric queries.
Refactored influxdb formulateQuery to base class with a couple of parameters to reflect the minor differences between metric and event queries.
Moved EventRepository CriteriaQueryBuilders back to the JDOEventRepository - they are not usable for the InfluxDB implementation.
Refactored some query methods (sum and mostRecentXXX) in JDOMetricsRepository to share common code.
Removed unused and unneeded API from MetricRepository interface.  Refactored common implementations to default interface methods.

Disabled (added @Ignore) annotation to 2 tests that fail if the InfluxDB service is not running. These succeeded with ObjectDB only because it runs completely embedded and starts automatically in referenced in code.

Added a simple test for the event query.  Fixed the order of expected, actual results in the assertions in TestInfluxRepository tests.