fs-1815 add influxdb url to platform.conf
Defined the fds_om_influxdb_url variable in the fds_config_defaults.j2 and if defined in the inventory, that is used for the value of the fds.om.influxdb.url setting in the generated platform.conf.  If fds_om_influxdb_url is not defined in the inventory, it generates a default URL using the om_ip variable.