fs-1676 - remove ObjectDB and JPA/JDO support
This is the nuclear option for fs-1676 and involves removing not only objectdb but all JPA/JDO references in the current code base as well.  One impact of this was in removing the hibernate dependency, we lost the Dom4J transitive dependency.  Given that Dom4J is effectively dead as an open source project, I chose to modify the two classes that used it to use the org.w3c.dom code that is part of the JDK.