fs-1863 handle lookup of metric by name or key
Replace the Metrics.byMetadataKey with a more generic lookup method that handles both the enum name and the key.  Modified FirebreakHelper to log debug messages on missing key (instead of warnings with a stack trace) that are ignored, and also addressed some of the many compiler warnings in that class.

Added filter to InfluxMetricRepository to filter out the InfluxDB sequence_number column from results instead of treating it like a VolumeDatapoint.