fs-1693 log files in bin directory
Log files were being created in the bin directory for the OM with a
default name/prefix of "log".  It turns out this was due to a log
message written in DeleteScheduler after starting a thread.  Since
DeleteScheduler was initialized in the constructor for the OM, this was
happening before the OM's log was fully initialized.  Moved the
initialization of the DeleteScheduler thread to occur later in the
startup process in OrchMgr::run

I modified the log initialization in fds_process to assert if the log is
not properly initialized in order to prevent components from trying to
use logging before it is ready and fully initialized.