Add Gson serialization adapter for VolumeSettings.
This adapter adds a CLASSNAME field to the Json string which is then used for the deserialization process.  It can be safely ignored by other clients of the json.  Extracted inner classes to address inner-class lookup problem in the deserialization process (also changed names to move the Type to the end of the name.  This just allows them to be grouped together in a file listing and has no other purpose).  Note that the CLASSNAME is redundant with the existing type field in the settings.  There may be a way to use that type field instead of adding CLASSNAME, but it will require more research.

Modified Size to remove the generic parameter and always store as a BigDecimal, which can handle any value we throw at it.  Added static factory methods that handle various primitive types.

Cleaned up the WeekDays class -- no need for it to be generic (ultimately it should be changed to make use of the existing iCalWeekdays values).
Copied some iCal related tests into new structure
Added v08 model GSON unit tests that do basic to/from json.
Fixed some existing unit tests that relied on old volume model.