More specific install versioning
In order for patch-level updates to versions to work with rolling deploy
we need to include the iteration of a version in the install directory
name. This basically means that a "version" becomes the full combination
of release date + git sha such as:

`2015.07.03_1-c9d64e4`

This allows for patches to be deployed & rolled back even within a
sprint version.