Use Firebreak Event for volume list query
This addresses one small part of the issue reported in fs-2660 Influx CPU usage when executing firebreak queries.  With this change we will query firebreak events from the event_db instead of processing 24 hours worth of metrics each time.  The fact is the generation of the event has already processed the metrics so there is no need to do it every time.  The primary changes specific to this are in ExternalModelConverter, FirebreakHelper, and InfluxEventRepository.

I refactored the Firebreak Event Cache out of the VolumeDatapointEntityPersistListener and moved it to FirebreakHelper (I have not yet taken the next step to actually use it elsewhere - the event query seems to return very quickly for the moment, perhaps because there are very few events).