Query date range fixes to help address influx query performance issues
Added static factory methods to DateRange for common usages: last24Hours, since, and between.

Added a TimeUnit to the DateRange class to indicate the unit that the range is defined in.  Defaults
to SECONDS in order to be consistent with UI marshalling since it does not currently pass in a unit (but
it is always SECONDS in the UI due to JavaScript representation as a 64-bit float meaning only 52 bits
are available for a large integer number)

Removed unused builder.

Added comment to InfluxDBReader interface to add more detail from the InfluxDB docs regarding the precision
parameter.  Specifically, for a query it indicates the precision of the time for the data returned.

Modified Influx repository to use the TimeUnit that is now in the data range, and attempt to be more consistent

In SystemHealthStatus (v07 and v08) only query the volume list from the server once.  Use new DateRange.last24Hours() api.