Fix for Bug 2930 : AddService returning API exception
The old code used a GetService call to verify whether an existing
service was being added. However, if the service does not exist, GetService
throws an exception. So now, the checks to verify addition of exisiting
service is done in send_add_service and if there is an error, it  is bubbled
back up to the Java API code.