fs-2942 replace OM config cache with xdi's config cache
Moved the XDI CachedConfiguration to the util.thrift package and modified OM to use it in place of the existing cache.  Modified OmConfigurationApi first check the cache for presence of item and only then refresh the cache.  Added platform.conf property to control the cache update interval when it validates the config version against the server and reloads the cache if necessary.