FS-2793 define FDS/XDI/OM JAVA_HOME vars
This defines new environment variables used to govern the specific JVM
used when starting the XDI or OM java processes.  This allows for the
JVM to be overridden as necessary in development, test, and potentially
support situations without requiring modifications to the system
fds-deps JRE installation.

By default, the FDS_JAVA_HOME variable will be set to
/opt/fds-deps/embedded/jre as is the default today.  XDI_JAVA_HOME and
OM_JAVA_HOME both default to ${FDS_JAVA_HOME}.  If FDS_JAVA_HOME is
overridden in the external environment, it will be used instead of the
default.

To provide more fine-grained control over the OM and XDI java processes,
the OM_JAVA_HOME or XDI_JAVA_HOME variables may each be set in the external
environment which will then override the java installation for the OM or
XDI processes, respectively.