FS-3004 fix memchk/release commands
the [no]memchk and debug|release commands were broken in the previous
commit where I moved some of the variable initialization so it happened
after reading the command-line arguments to detect the --fds-root
argument, if provided.  This moves some of that init back to the top of
the file and then streamlines some of the command-line argument
processing.

This also removes the -i|--from-install-dir option which appears to have
been broken for as far back in the file history as I looked.  It
incorrectly sets the etc directory so could not have worked unless the
etc directory lived in a different location at some point in time.

Note that the --fds-root option will currently only work if running with
a clean redis database or distinct configurations (i.e. not using the
default symlinks from the build/dev_make_install.sh setup).