Address review comments on OM Port configuration
All of the service ports except for the OM are mapped based on offsets from the Platform Port.  Although the default OM port happens to be an offset from the default platform port as well, in fact, the OM port is based on the configuration of fds.common.om_port.  This commit attempts to address not only the one review comment, but all of the places in the core code where are initializing an OM port value.  I found that we were failing to start properly when attempting to run the OM on a non-default port and I believe this addresses all of those places.

I added logging that will log the actual ip and port it is attempting to connect to which will hopefully aid future troubleshooting.

This also adds a configuration for the OM Service Proxy Port offset, used when the java service proxy is enabled.

Note that this does not address tools that may access the default OM port, such as fdstrace.