fs-2780 ensure JVM heap/crash logs in /corefiles
Updates the runtime initialization of the HeapDumpOnOutOfMemoryError and
HeapDumpPath settings ensuring that it is enabled and defaulting the
path to write to /corefiles.  It is still possible to override the
location by explicitly including it on the JVM command-line.

The JVM -XX:ErrorFile setting can not be modified at runtime, so it is
added to the OM and XDI JAVA_OPTS settings.

This also includes changes to the log configuration that ensures that
logging is configured with appropriate defaults if not explicitly
specified in the default *_JAVA_OPTS settings that include
-DconfigurationFile=${fds_root}/etc/log4j2.xml

These default settings would take affect if the property is not set, or
if the log4j2.xml file is not found at startup time.  The most likely
scenario where that would happen is debugging in an IDE.