fs-3131 fix log4j cfg properties
Modifies the log4j2.xml configuration template to define a default
fds_root value that is expected to be overridden by the environment
setting (via formation_env etc).  Modified the fds.service.name to
use the "sys" context where it will first look for the definition in the
system properties and if not defined, will use "default".  This isn't a
perfect solution, but coupled with the explicit setting of
fds.service.name system property in the programmatic configuration in
Configuration.java, will address the issue we are seeing.

There is still a chicken-and-egg situation with the log4j configuration.
If -DconfigurationFile is specified, but fds.service.name is not, the
log file will be created as "default.log" which is not what we want.

In addition to the fixes for the log file naming and property
resolution, the also changes the log rollover policies.  I have removed
the date/timestamp from the file pattern and the associated
TimeBasedTriggeringPolicy.  This was necessary because currently Log4J
has bugs where  the DefaultRolloverStrategy file limit is unable to
remove old log files when using a timestamp
(https://issues.apache.org/jira/browse/LOG4J2-435).