Remove 10 argument limit to process execution
Adds checking to validate that the number of args passed in matches the
expected args and does not exceed the new max (256).  On the fds root
path modification for configdb use, ensure that the path is actually
modified and there is space to add the null termination (otherwise
potentially overwriting memory).