FS-3585:
There is a broken pipe being reported, so service state is being set to not running.
There is also a case where svc info incarnation number is set to zero, which was being processed, but ignored because 'zero' isn't a valid incarnation number.

But looks like it should be treated as valid.

Will put code fix in place to make zero acceptable incarnation number.