FS-xxx: SM SuperBlock was not properly packed which...
was causing portions of it to be uninitialized and contain random values. The main
reason was that the fields in the superblock were not POD as they defined their own
desctructors and assignment operators...etc.

This fix should not be merged until it is either verified that it does not break
backwards compatibility after an upgrade, or made to work so that it does.