FS-2748 attempt to workaround influxdb cpu/mem issues by serializing reads and/or writes.
Introduces three flags to the platform.conf to configure the InfluxDBConnection to serialize reads, writes, or both
    fds.om.influxdb.serialize_reads
    fds.om.influxdb.serialize_writes
    fds.om.influxdb.serialize_all

If serialize_all is true, InfluxDBConnection will create InfluxDBReader and InfluxDBWriter that uses a shared lock to ensure all reads and writes are serialized.

Otherwise, if
  - serialize_reads=true, all read queries will be serialized.
  - serialize_writes=true, all write queries will be serialized.

If serialize_all is false and both reads/writes are true, the reads will be serialized against other reads; and writes will similarly be serialized against other writes, but not against the other.