Adding fdsconsole flag to inventory files
This commit is to add the install_fdsconsole flag to inventory files
known to be used to deploy and run system tests on a full cluster. The
fdsconsole tool has been removed for standard deployments, but is still
needed when running system tests.

If you are running such tests on a deployed cluster, and it is not able
to run due to not finding fdsconsole, you will need to include this flag
in your inventory file, and redeploy.

Eventually, System Test's dependency on fdsconsole should be removed
allowing for the complete deprication of fdsconsole.