Direct references to RequestLog.newRequestLogger no longer necessary
Changes were made to the RequestLog wrappers such that now it is able to
directly wrap either a generic HttpServletRequest or a Jetty-specific
request.  The dispatcher was modified to always wrap the Jetty request
so by the time it makes it to the handler, it is already wrapped.
Although the newRequestLogger(request) impl takes care to return an
existing wrapper if present, I felt it was cleaner to remove these
references. 

The only significant change in any of these files is the removal of the
call to RequestLog.newRequestLogger call.  Unfortunately a formatter
setting was enabled in my IDE environment that resulted in the
reformatting of these files.  I have decided to leave the format changes
in place since they are more consistent with other code format (whereas
the previous format was not) and this is an otherwise simple change.