Use a single StatStreamRegistration for the OM
Previously we were using a stat stream registration per-volume.  Due to
a bug in the DM it was sending stats for all volumes in each stat stream
registration, resulting in the processing of a lot of extra duplicate
data.  Somewhere along the way, the DM also started supporting the
ability to send all volume stats to a StatStreamRegistration that had an
empty volume name list.  This is what we have wanted for the OM all
along - the ability to have a single StatStreamRegistration that
captures data for all volumes.

The StatStreamRegistrationHandler was modified to check for existing
per-volume registrations and remove any with a destination of the OM URL
and then make sure that it has a empty volume list registration.

This also adds detailed trace logging to the AM/XDI StatisticPublisher
that dumps all of the stats it receives to the log file before sending
it to the destination url (including OM).