FS-4920 Memoization Cache for frequently computed values.
The Memoizer utilizes the Java 8 Supplier and Function to implement
memoization for potentially long calculations that are frequently
requested.  It supports an optional expiration (ttl) that after which
point the cached value is discarded and it is recomputed.

In QueryHelper, the CapacityToFull calculation is wrapped in a function
and memoized through the memoizer.