FS-4865 - fix error message reporting from OM REST error handler
Modify the OM error handler to add information to the errors reported,
including exception type and if a ApiException, the error code.  

Modify the v08 OMConfigServiceRestClientv08Impl to read the new error
response information.

Modified the OMConfigException to extend ApiException which is the
expected exception type of config svc interfaces.  This avoids
UndeclaredThrowableExceptions and also includes the error code
information that the XDI S3 connector depends on for interpreting errors
properly.