FS-4865 - remove duplicate creation go the _vol1 bucket
When using the v07 REST API, the creation of a bucket that already
exists does not result in an error - it returned the existing bucket.
When we use the v08 API it returns an error, which was causing this test
case that creates the _vol1 bucket multiple times to fail.