FS-5004: Try and fix MP relying on volume attach.
This change pushes the required information into the Block layer if the connector
has it, as the iSCSI connector does. This should allow a client to attach and perform
any non-IO (read, write) commands. Hopefully this provides MP support without providing
read from anywhere support as VolumeGrouping does not currently.