FS-5400 add a log4j config file for the smoketest.
This file can be modified to add/modify/remove loggers for debugging the
smoke test as necessary.  It provides a minimal defualt configuration
equivalent to the programmatic configuration previously hard-coded in
SmokeTestRunner.  To enable debug logging, edit the file and change the
levels to DEBUG, or use a command like

 sed -i -r 's/level="(\w+)"/level="DEBUG"/g' smoketest-log4j2.xml

to change all of the levels.